# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FacturaloYa",
        "description": "Sistema de facturacion en linea. Con carrito de compras",
        "contact": {
            "email": "neisserrey@upeu.edu.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://facturaloya.test/api",
            "description": "Documentacion de la Facturacion en Linea"
        },
        {
            "url": "http://facturaloya.test/api/"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Iniciar sesion para obtener un token de acceso.",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciales del usuario (username, password)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Datos invalidos!"
                    },
                    "201": {
                        "description": "Successful pues :3"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/categorias": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Mostrar categorias",
                "operationId": "App\\Http\\Controllers\\Categoria\\CategoriaController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de categorias. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "categorias"
                ],
                "summary": "Agregar una nueva categoria.",
                "operationId": "store",
                "requestBody": {
                    "description": "Categoria a ser creada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/categorias/{categoriaId}": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Buscar categoria por ID",
                "description": "Retorna una sola categoria",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "description": "ID de la categoria a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "Categoria no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "categorias"
                ],
                "summary": "Actualizar una categoria existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "description": "ID de la categoria a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Categoria a ser actualizada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Categoria actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categorias"
                ],
                "summary": "Elimina una categoria",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "description": "Id de la categoria a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Categoria no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "403": {
                        "description": "Tiene productos relacionados"
                    },
                    "200": {
                        "description": "Categoria eliminada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/clientes": {
            "get": {
                "tags": [
                    "clientes"
                ],
                "summary": "Mostrar clientes",
                "operationId": "App\\Http\\Controllers\\Cliente\\ClienteController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de clientes. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "clientes"
                ],
                "summary": "Agrega un nuevo Cliente.",
                "operationId": "store",
                "requestBody": {
                    "description": "Cliente a ser agregado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/clientes/{clienteId}": {
            "get": {
                "tags": [
                    "clientes"
                ],
                "summary": "Buscar Cliente por ID",
                "description": "Retorna un solo Cliente",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "clienteId",
                        "in": "path",
                        "description": "ID del cliente a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "Cliente no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "clientes"
                ],
                "summary": "Actualizar un cliente existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "clienteId",
                        "in": "path",
                        "description": "ID del cliente a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Cliente a ser actualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Cliente actualizado correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "clientes"
                ],
                "summary": "Elimina un Cliente",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "clienteId",
                        "in": "path",
                        "description": "Id del cliente a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Cliente no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "403": {
                        "description": "Tiene productos relacionados"
                    },
                    "200": {
                        "description": "Cliente eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/tipo_documentos": {
            "get": {
                "tags": [
                    "tipo_documentos"
                ],
                "summary": "Mostrar tipo de documentos",
                "operationId": "App\\Http\\Controllers\\TipoDocumento\\TipoDocumentoController::index",
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Lista de tipo de documentos. Correcto"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tipo_documentos"
                ],
                "summary": "Agregar un nuevo tipo de documento(DNI,CARNET DE EXTRANJERIA, CEDULA,ETC).",
                "operationId": "store",
                "requestBody": {
                    "description": "TipoDocumento a ser creado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoDocumento"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "201": {
                        "description": "Creado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/tipo_documentos/{tipo_documento_id}": {
            "get": {
                "tags": [
                    "tipo_documentos"
                ],
                "summary": "Buscar tipo de documentos por ID",
                "description": "Retorna un solo tipo de documento",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "tipo_documento_id",
                        "in": "path",
                        "description": "ID del tipo de documento a retornar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoDocumento"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "404": {
                        "description": "TIpo de documento no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/tipo_documentos/{tipoDocumentoId}": {
            "put": {
                "tags": [
                    "tipo_documentos"
                ],
                "summary": "Actualizar un tipo de documento existente",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "tipoDocumentoId",
                        "in": "path",
                        "description": "ID del documento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tipo de documento a ser actualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoDocumento"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Tipo de documento no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Tipo de documento actualizada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "tipo_documentos"
                ],
                "summary": "Elimina una tipo de documento",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "tipoDocumentoId",
                        "in": "path",
                        "description": "Id del tipo de documento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Tipo de documento no encontrada"
                    },
                    "401": {
                        "description": "Inserta tu token pues hermano!"
                    },
                    "200": {
                        "description": "Tipo de documento eliminada correctamente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Categoria": {
                "properties": {
                    "nombre": {
                        "description": "Nombre del producto",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripcion del producto",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Cliente": {
                "properties": {
                    "nombres": {
                        "description": "Nombres del cliente",
                        "type": "string"
                    },
                    "apellidos": {
                        "description": "Apellidos del cliente",
                        "type": "string"
                    },
                    "correo": {
                        "description": "correo del cliente",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "telefono del cliente",
                        "type": "int"
                    },
                    "direccion": {
                        "description": "direccion del cliente",
                        "type": "string"
                    },
                    "numero_documento": {
                        "description": "numero_documento del cliente",
                        "type": "int"
                    },
                    "tipo_documento_id": {
                        "description": "tipo_documento_id del cliente",
                        "type": "int"
                    }
                },
                "type": "object"
            },
            "DetalleVenta": {
                "properties": {
                    "cantidad": {
                        "description": "Cantidad del detalle de venta",
                        "type": "int"
                    },
                    "precio_unitario": {
                        "description": "Precio unitario del detalle de la venta",
                        "type": "int"
                    },
                    "precio_total": {
                        "description": "Precio del detalle de la venta",
                        "type": "int"
                    },
                    "producto_id": {
                        "description": "Id del producto",
                        "type": "int"
                    },
                    "venta_id": {
                        "description": "Id de la venta a realizar",
                        "type": "int"
                    }
                },
                "type": "object"
            },
            "TipoDocumento": {
                "properties": {
                    "nombre": {
                        "description": "Nombre del tipo de documento",
                        "type": "string"
                    },
                    "tabla": {
                        "description": "tabla del tipo de documento",
                        "type": "string"
                    },
                    "tamaño": {
                        "description": "tamaño del tipo de documento",
                        "type": "int"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "username": {
                        "description": "username del usuario",
                        "type": "string"
                    },
                    "password": {
                        "description": "password del usuario",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}